Changed clock on v65C02 to 64-bit, more GUI refactoring

+++ src/v65c02.h
@@ -37,7 +37,8 @@ struct V65C02REGS
 	uint8 a;						// 65C02 A register
 	uint8 x;						// 65C02 X index register
 	uint8 y;						// 65C02 Y register
-	uint32 clock;					// 65C02 clock (@ 1 MHz, wraps at 71.5 minutes)
+//	uint32 clock;					// 65C02 clock (@ 1 MHz, wraps at 71.5 minutes)
+	uint64 clock;					// 65C02 clock (@ 1 MHz, wraps at 570842 years)
 	uint8 (* RdMem)(uint16);		// Address of BYTE read routine
 	void (* WrMem)(uint16, uint8);	// Address of BYTE write routine
 	uint16 cpuFlags;				// v65C02 IRQ/RESET flags
@@ -50,6 +51,6 @@ extern bool dumpDis;
 // Function prototypes
 
 void Execute65C02(V65C02REGS *, uint32);    	// Function to execute 65C02 instructions
-uint32 GetCurrentV65C02Clock(void);				// Get the clock of the currently executing CPU
+uint64 GetCurrentV65C02Clock(void);				// Get the clock of the currently executing CPU
 
 #endif	// __V65C02_H__
